==> 2021 - IJERPH - SHARE Migrants/03 - Recodings Countries.R <==
# http://unstats.un.org/unsd/methods/m49/m49alpha.htm

country_code <- read.table(
  text = "Afghanistan_004_AFG
Åland Islands_248_ALA
Albania_008_ALB
Algeria_012_DZA
American Samoa_016_ASM
Andorra_020_AND
Angola_024_AGO
Anguilla_660_AIA
Antarctica_010_ATA
Antigua and Barbuda_028_ATG
Argentina_032_ARG
Armenia_051_ARM
Aruba_533_ABW
Australia_036_AUS
Austria_040_AUT
Azerbaijan_031_AZE
Bahamas_044_BHS
Bahrain_048_BHR
Bangladesh_050_BGD
Barbados_052_BRB
Belarus_112_BLR
Belgium_056_BEL
Belize_084_BLZ
Benin_204_BEN
Bermuda_060_BMU
Bhutan_064_BTN
Bolivia (Plurinational State of)_068_BOL
Bonaire, Sint Eustatius and Saba_535_BES
Bosnia and Herzegovina_070_BIH
Botswana_072_BWA
Bouvet Island_074_BVT
Brazil_076_BRA
British Indian Ocean Territory_086_IOT
British Virgin Islands_092_VGB
Brunei Darussalam_096_BRN
Bulgaria_100_BGR
Burkina Faso_854_BFA
Burundi_108_BDI
Cabo Verde_132_CPV
Cambodia_116_KHM
Cameroon_120_CMR
Canada_124_CAN
Cayman Islands_136_CYM
Central African Republic_140_CAF
Chad_148_TCD
Chile_152_CHL
China_156_CHN
China, Hong Kong Special Administrative Region_344_HKG
China, Macao Special Administrative Region_446_MAC
Christmas Island_162_CXR
Cocos (Keeling) Islands_166_CCK
Colombia_170_COL
Comoros_174_COM
Congo_178_COG
Cook Islands_184_COK
Costa Rica_188_CRI
Côte d’Ivoire_384_CIV
Croatia_191_HRV
Cuba_192_CUB
Curaçao_531_CUW
Cyprus_196_CYP
Czechia_203_CZE
Democratic People's Republic of Korea_408_PRK
Democratic Republic of the Congo_180_COD
Denmark_208_DNK
Djibouti_262_DJI
Dominica_212_DMA
Dominican Republic_214_DOM
Ecuador_218_ECU
Egypt_818_EGY
El Salvador_222_SLV
Equatorial Guinea_226_GNQ
Eritrea_232_ERI
Estonia_233_EST
Eswatini_748_SWZ
Ethiopia_231_ETH
Falkland Islands (Malvinas)_238_FLK
Faroe Islands_234_FRO
Fiji_242_FJI
Finland_246_FIN
France_250_FRA
French Guiana_254_GUF
French Polynesia_258_PYF
French Southern Territories_260_ATF
Gabon_266_GAB
Gambia_270_GMB
Georgia_268_GEO
Germany_276_DEU
Ghana_288_GHA
Gibraltar_292_GIB
Greece_300_GRC
Greenland_304_GRL
Grenada_308_GRD
Guadeloupe_312_GLP
Guam_316_GUM
Guatemala_320_GTM
Guernsey_831_GGY
Guinea_324_GIN
Guinea-Bissau_624_GNB
Guyana_328_GUY
Haiti_332_HTI
Heard Island and McDonald Islands_334_HMD
Holy See_336_VAT
Honduras_340_HND
Hungary_348_HUN
Iceland_352_ISL
India_356_IND
Indonesia_360_IDN
Iran (Islamic Republic of)_364_IRN
Iraq_368_IRQ
Ireland_372_IRL
Isle of Man_833_IMN
Israel_376_ISR
Italy_380_ITA
Jamaica_388_JAM
Japan_392_JPN
Jersey_832_JEY
Jordan_400_JOR
Kazakhstan_398_KAZ
Kenya_404_KEN
Kiribati_296_KIR
Kuwait_414_KWT
Kyrgyzstan_417_KGZ
Lao People's Democratic Republic_418_LAO
Latvia_428_LVA
Lebanon_422_LBN
Lesotho_426_LSO
Liberia_430_LBR
Libya_434_LBY
Liechtenstein_438_LIE
Lithuania_440_LTU
Luxembourg_442_LUX
Madagascar_450_MDG
Malawi_454_MWI
Malaysia_458_MYS
Maldives_462_MDV
Mali_466_MLI
Malta_470_MLT
Marshall Islands_584_MHL
Martinique_474_MTQ
Mauritania_478_MRT
Mauritius_480_MUS
Mayotte_175_MYT
Mexico_484_MEX
Micronesia (Federated States of)_583_FSM
Monaco_492_MCO
Mongolia_496_MNG
Montenegro_499_MNE
Montserrat_500_MSR
Morocco_504_MAR
Mozambique_508_MOZ
Myanmar_104_MMR
Namibia_516_NAM
Nauru_520_NRU
Nepal_524_NPL
Netherlands_528_NLD
New Caledonia_540_NCL
New Zealand_554_NZL
Nicaragua_558_NIC
Niger_562_NER
Nigeria_566_NGA
Niue_570_NIU
Norfolk Island_574_NFK
North Macedonia_807_MKD
Northern Mariana Islands_580_MNP
Norway_578_NOR
Oman_512_OMN
Pakistan_586_PAK
Palau_585_PLW
Panama_591_PAN
Papua New Guinea_598_PNG
Paraguay_600_PRY
Peru_604_PER
Philippines_608_PHL
Pitcairn_612_PCN
Poland_616_POL
Portugal_620_PRT
Puerto Rico_630_PRI
Qatar_634_QAT
Republic of Korea_410_KOR
Republic of Moldova_498_MDA
Réunion_638_REU
Romania_642_ROU
Russian Federation_643_RUS
Rwanda_646_RWA
Saint Barthélemy_652_BLM
Saint Helena_654_SHN
Saint Kitts and Nevis_659_KNA
Saint Lucia_662_LCA
Saint Martin (French Part)_663_MAF
Saint Pierre and Miquelon_666_SPM
Saint Vincent and the Grenadines_670_VCT
Samoa_882_WSM
San Marino_674_SMR
Sao Tome and Principe_678_STP
Sark_680_SARK
Saudi Arabia_682_SAU
Senegal_686_SEN
Serbia_688_SRB
Seychelles_690_SYC
Sierra Leone_694_SLE
Singapore_702_SGP
Sint Maarten (Dutch part)_534_SXM
Slovakia_703_SVK
Slovenia_705_SVN
Solomon Islands_090_SLB
Somalia_706_SOM
South Africa_710_ZAF
South Georgia and the South Sandwich Islands_239_SGS
South Sudan_728_SSD
Spain_724_ESP
Sri Lanka_144_LKA
State of Palestine_275_PSE
Sudan_729_SDN
Suriname_740_SUR
Svalbard and Jan Mayen Islands_744_SJM
Sweden_752_SWE
Switzerland_756_CHE
Syrian Arab Republic_760_SYR
Tajikistan_762_TJK
Thailand_764_THA
Timor-Leste_626_TLS
Togo_768_TGO
Tokelau_772_TKL
Tonga_776_TON
Trinidad and Tobago_780_TTO
Tunisia_788_TUN
Turkey_792_TUR
Turkmenistan_795_TKM
Turks and Caicos Islands_796_TCA
Tuvalu_798_TUV
Uganda_800_UGA
Ukraine_804_UKR
United Arab Emirates_784_ARE
United Kingdom of Great Britain and Northern Ireland_826_GBR
United Republic of Tanzania_834_TZA
United States Minor Outlying Islands_581_UMI
United States of America_840_USA
United States Virgin Islands_850_VIR
Uruguay_858_URY
Uzbekistan_860_UZB
Vanuatu_548_VUT
Venezuela (Bolivarian Republic of)_862_VEN
Viet Nam_704_VNM
Wallis and Futuna Islands_876_WLF
Western Sahara_732_ESH
Yemen_887_YEM
Zambia_894_ZMB
Zimbabwe_716_ZWE",
  quote = "",
  sep = "_"
)
colnames(country_code) <- c("country", "id", "code")


share$country_born_string <- NA
share$country_born_code <- NA
for (i in na.omit(unique(share$country_born))) {
  if (i %in% country_code$id) {
    share$country_born_string[which(share$country_born == i)] <- country_code$country[country_code$id == i]
    share$country_born_code[which(share$country_born == i)] <- country_code$code[country_code$id == i]
  } else {
    warning(paste0("Country code ", i, " not in data."), call. = FALSE)
  }
}

share$country_of_birth <- share$country_born
share$country_of_birth <- set_labels(
  share$country_of_birth,
  labels = setNames(country_code$id, country_code$country)
)
